change int to double in closerToOrigin and Runner

+++ src/task7/ClosestPointLogic.java
@@ -1,7 +1,7 @@
 package task7;
 
 public class ClosestPointLogic {
-    public String closerToOrigin(int x1, int y1, int x2, int y2) {
+    public String closerToOrigin(double x1, double y1, double x2, double y2) {
         double distance1 = Math.sqrt(Math.pow(x1, 2) + Math.pow(y1, 2));
         double distance2 = Math.sqrt(Math.pow(x2, 2) + Math.pow(y2, 2));
